backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
python-multipart
pdfplumber
spacy
scikit-learn
requests
python-dotenv
slowapi